refactor: Move graph ql to another file
