+++ app/src/main/java/com/example/demopokekotlin/Pokemon.kt
@@ -1,44 +1,8 @@
 package com.example.demopokekotlin
 
 import android.util.Log
-import okhttp3.MediaType.Companion.toMediaType
-import okhttp3.OkHttpClient
-import okhttp3.Request
-import okhttp3.RequestBody.Companion.toRequestBody
 import org.json.JSONArray
 import org.json.JSONObject
-
-class GraphQLApi(private val endpoint: String) {
-    fun queryJson(query: String): String? {
-        val http = OkHttpClient()
-
-        val contentType = "application/json; charset=utf-8".toMediaType()
-        val json = JSONObject()
-
-        try {
-            json.put("query", query)
-            json.put("variables", JSONObject())
-            json.put("operationName", "pokemons_list")
-        } catch (e: Exception) {
-            Log.w("POKEMONS", "Invalid query: ${e.message}.")
-        }
-
-        val body = json.toString().toRequestBody(contentType)
-        val request = Request.Builder()
-            .url(endpoint)
-            .post(body)
-            .build()
-
-        val response = http.newCall(request).execute()
-
-        if (response.body == null) {
-            Log.w("POKEMONS", "Failed to load pokemons: code ${response.code}.")
-            return null
-        }
-
-        return response.body!!.string()
-    }
-}
 
 // TODO: Implement factory pattern
 class Pokemon {
@@ -53,18 +17,24 @@ class Pokemon {
             val pokemons = mutableListOf<Pokemon>()
             val client = GraphQLApi("https://beta.pokeapi.co/graphql/v1beta/")
 
-            val query = "query pokemons_list {\n" +
-                    "  pokemon_v2_pokemon {\n" +
-                    "    id\n" +
-                    "    name\n" +
-                    "    pokemon_v2_pokemontypes{\n" +
-                    "      pokemon_v2_type {\n" +
-                    "        id\n" +
-                    "        name\n" +
-                    "      }\n" +
-                    "    }\n" +
-                    "  }\n" +
-                    "}\n"
+            /*val reader = FileReader("pokemons.gql")
+            val query = reader.readText()
+            reader.close()*/
+
+            val query = """
+                query pokemons_list {
+                    pokemon_v2_pokemon {
+                        id
+                        name
+                        pokemon_v2_pokemontypes {
+                            pokemon_v2_type {
+                                id
+                                name
+                            }
+                        }
+                    }
+                }
+            """.trimIndent()
 
             failure.name = "Failed to load"
             failure.type = "Rip bozo"
@@ -83,17 +53,7 @@ class Pokemon {
             }
 
             try {
-                val dataObject = json.getJSONObject("data")
-                val pokemonsArray = dataObject.getJSONArray("pokemon_v2_pokemon")
-
-                for (i in 0 until pokemonsArray.length()) {
-                    try {
-                        parsePokemon(pokemonsArray, i, pokemons)
-                    } catch(e: Exception) {
-                        Log.w("POKEMONS", "Failed to load pokemon $i: ${e.message}.")
-                        //pokemons.add(failure)
-                    }
-                }
+                parsePokemons(json, pokemons)
             } catch(e: Exception) {
                 Log.w("POKEMONS", "Failed to load pokemons: ${e.message}.")
                 return listOf()
@@ -104,6 +64,20 @@ class Pokemon {
 
         fun getDetailsOf() {
 
+        }
+
+        private fun parsePokemons(json: JSONObject, pokemons: MutableList<Pokemon>) {
+            val dataObject = json.getJSONObject("data")
+            val pokemonsArray = dataObject.getJSONArray("pokemon_v2_pokemon")
+
+            for (i in 0 until pokemonsArray.length()) {
+                try {
+                    parsePokemon(pokemonsArray, i, pokemons)
+                } catch(e: Exception) {
+                    Log.w("POKEMONS", "Failed to load pokemon $i: ${e.message}.")
+                    //pokemons.add(failure)
+                }
+            }
         }
 
         private fun parsePokemon(
